requirements.txt
transformers
requests
datasets
torch
numpy
vlite2
python-dotenv
openai